Code coverage generation and upload to codecov.io
This commit adds code coverage generation when building with GCC or Clang
on Unix-based systems.

Enable it by setting Mapper_COVERAGE_BUILD to ON (the default when
Mapper_DEVELOPMENT_BUILD is set).

The main interesting target is coverage_html, which will run tests and
generate a HTML coverage report in <builddir>/coverage/html. Additionally
there is a coverage_codecov target that will upload coverage data to
codecov.io (this needs to either be run on a CI or with an explicit repo
token to work).